requirements.txt
Pillow
google-generativeai
streamlit
IPython